+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.19.0] - 2021-10-20
+
+### Added
+
+-   Added the ability to choose own console log timezone (Default: Host timezone)
+-   Added `admins` slash command (Default: Only admins) which displays all admins (on a specified server) with their playtime and use any past days (Default: 14 days)
+
 ## [1.18.7] - 2021-10-09
 
 ### Fixed